requirements.txt
aiohttp~=3.8.4
aiofiles~=23.1.0
aiohttp_session~=2.12.0
cryptography~=41.0.1
Pillow~=9.5.0
torch~=2.0.1
torchvision~=0.15.2
numpy~=1.24.3
pycocotools~=2.0.6
PyYAML~=6.0
marshmallow~=3.19.0